+++ examples/tools/scene_viewer/mod.rs
@@ -1,0 +1,2 @@
+pub mod camera_controller_plugin;
+pub mod scene_viewer_plugin;
